+++ app/src/main/java/com/nbhysj/coupon/model/response/GroupMchDetailsResponse.java
@@ -351,7 +351,7 @@ public class GroupMchDetailsResponse implements Serializable {
 
         private float commentScore;
 
-        private int consumePrice;
+        private double consumePrice;
 
         private String mchRanking;
 
@@ -431,11 +431,11 @@ public class GroupMchDetailsResponse implements Serializable {
             this.photo = photo;
         }
 
-        public int getConsumePrice() {
+        public double getConsumePrice() {
             return consumePrice;
         }
 
-        public void setConsumePrice(int consumePrice) {
+        public void setConsumePrice(double consumePrice) {
             this.consumePrice = consumePrice;
         }
 
@@ -1014,7 +1014,7 @@ public class GroupMchDetailsResponse implements Serializable {
 
         private int score;
 
-        private int price;
+        private double price;
 
         private String photo;
 
@@ -1068,11 +1068,11 @@ public class GroupMchDetailsResponse implements Serializable {
             this.score = score;
         }
 
-        public int getPrice() {
+        public double getPrice() {
             return price;
         }
 
-        public void setPrice(int price) {
+        public void setPrice(double price) {
             this.price = price;
         }
