+++ app/src/main/java/com/nbhysj/coupon/ui/ScenicSpotDetailActivity.java
@@ -419,7 +419,7 @@ public class ScenicSpotDetailActivity extends BaseActivity<ScenicSpotPresenter, 
                     int level = mchDetailsEntity.getLevel();
                     //评价
                     int commentNum = mchDetailsEntity.getCommentNum();
-                    int mConsumePrice = mchDetailsEntity.getConsumePrice();
+                    double mConsumePrice = mchDetailsEntity.getConsumePrice();
                     float mCommentScore = mchDetailsEntity.getCommentScore();
                     List<MchDetailsResponse.TagsEntity> tagsEntityList = mchDetailsEntity.getTags();
                     mTvScenicSpotName.setText(mchName + "(" + level + "A)");
